+++ checkout-ui-custom/src/templates/MixedProducts.js
@@ -1,5 +1,6 @@
 const MixedProducts = (message) => `
   <div id="tfg-custom-mixed-msg" class="tfg-custom-msg">
+    <p class="tfg-custom-icon"></p>
     <p class="tfg-custom-text">${message}</p>
   </div>
 `;