+++ checkout-ui-custom/src/templates/TVIDForm.js
@@ -3,7 +3,7 @@ const TVIDForm = () => `
         <p class="tfg-custom-title">TV licence required</p>
         <p class="tfg-custom-subtitle">Please provide your ID number to validate your TV Licence.</p>
         <p class="input tfg-custom-input tfg-tv-licence">
-            <label>ID Number</label>
+            <label>SA ID number</label>
             <input id="tfg-tv-licence" type="text" class="input-xlarge tfg-input">
         </p>
     </div>